clk: tegra: Update PLLM handling
PLLM is fixed for Tegra30 up through Tegra114. Starting with Tegra124
PLLM can change rate. Mark PLLM as TEGRA_PLL_FIXED for the generations
where it should be. Modify the check in clk_pll_round_rate and
clk_pll_recalc_rate to allow for the non-fixed version to
return the correct rate.

Note that clk-tegra20 is not touched. This is because PLL_M is not
distinguished in that driver, and adding either the PLL_M or FIXED_RATE
flags will cause potential problems.

PLLM never supported dynamic ramping. On T20/T30, there is no dynamic
ramping al all, and on T114/T124/T132, only PLLX and PLLC support
dynamic ramping, so we can go ahead and remove the specialized pllm_ops.

Signed-off-by: Danny Huang <dahuang@nvidia.com>
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>