WIP: drm/tegra: dc: Use direct window register accesses
Tegra supports two paths to address window registers: an indirect path
that requires selecting the active window for register operations along
with fixed register offsets. The direct path uses separate offsets for
window registers and does not require selecting the active window.

The indirect path has the drawback that access to window registers needs
to be serialized to ensure concurrent updates of multiple windows don't
stump on each other's toes. The direct access path does not have this
drawback and therefore simplifies window updates. It is also the method
recommended by the TRM, so implement it instead of the indirect path.

TODO: properly dump per-window registers

Signed-off-by: Thierry Reding <treding@nvidia.com>