drm/edid: Implement a helper to get the EDID
Querying the EDID is a rather common operation for display drivers that
need to selectively enable support for a feature advertised in one of
the EDID extension blocks. Introduce helpers that allow the EDID blob
property to be read while taking care of proper locking and reference
counting to ensure the data doesn't go away all of a sudden.

Signed-off-by: Thierry Reding <treding@nvidia.com>