arm64: tegra: Add NVIDIA P2597 base board support
The NVIDIA P2597 base board is a carrier board for the P2180 compute
module. I provides a RJ45 connector routed to the P2180 module's USB 3.0
network adapter, as well as a slew of other connectors such as SATA, USB
3.0, HDMI, JTAG and PCIe, among others. It also has dedicated connectors
to attach display and camera modules. A full-size SD slot is provided to
extend storage beyond the 16 GiB of eMMC found on the P2180.

Signed-off-by: Thierry Reding <treding@nvidia.com>