arm64: tegra: Add NVIDIA P2180 compute module support
The NVIDIA P2180 is a compute module used in several reference designs.
It features a Tegra210 SoC, 4 GiB of LPDDR4 RAM, 16 GiB eMMC and various
other essentials. It is typically connected to some base board (such as
the P2597) that provides the connectors needed to hook it up to the
outside world.

Signed-off-by: Thierry Reding <treding@nvidia.com>