PCI: tegra: Implement ->add_bus() callback
The configuration space mapping on Tegra is somewhat special, and in
order to avoid wasting virtual address space the configuration space
for each bus needs to be stitched together from several blocks which
form a single continuous virtual address range for accessors.

Currently the configuration space is mapped upon the first access to
one of its registers. However, the mapping operation may sleep under
certain circumstances, so doing it from the configuration space
accessors (they are protected by a spin lock) will trigger a warning.

To avoid the warning, use the ->add_bus() callback to perform the
mapping at enumeration time when the operation is allowed to sleep.

Signed-off-by: Thierry Reding <treding@nvidia.com>